FINALLY found an ideal method of formatting the flash cards in XAML
After debating whether to use stack layout or relative layout, find some good discussions and explinations of things I need to do that can be done will with grid layout. Started laying out the card elements in this format. Still needs lots of refining, but the base of the layout is there.