Flash card video is becoming more user friendly.
Still need to handle all kinds of edge cases for what to do with the video when a user starts/stops it, navigates away from it, flips the card during various playback scenarios, etc. Also, currently when user clicks the video button, a new video player is added to the layout. Need to fix it so that rather than adding a video, it just assigns the corrrect video path to the already existing video player. This will prevent duplicate video players from cluttering up the page (in both the foreground and background).