Minor tweaks to video and screen sizes
The image for the flash card was appearing too high on an iPhone with a taller screen. It was determined to alter the VerticalOptions property of the photo, and to then make sure a larger photo is being used as an asset to ensure proper placement of all the elements.

Also tried to tweak/improve the handling of video in the flash cards, but I don't think I made much of a difference.