The app will now remember the same list across all Pages
Tweaks were made so that when a user turns a card on or off (for use in a custom flash card deck), the app will now remember the status of each card. However, this isn't remembered when the app is closed, so this data needs to be saved more deeply.